+++ src/main/java/cn/malgo/annotation/dao/AnTypeRepository.java
@@ -4,6 +4,5 @@ import cn.malgo.annotation.entity.AnType;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
-/** Created by cjl on 2018/6/5. */
 public interface AnTypeRepository
-    extends JpaRepository<AnType, Integer>, JpaSpecificationExecutor {}
+    extends JpaRepository<AnType, Long>, JpaSpecificationExecutor {}
