+++ src/main/java/cn/malgo/annotation/dao/AnnotationTaskBlockRepository.java
@@ -3,7 +3,6 @@ package cn.malgo.annotation.dao;
 import cn.malgo.annotation.entity.AnnotationTaskBlock;
 import cn.malgo.annotation.enums.AnnotationTaskState;
 import cn.malgo.annotation.enums.AnnotationTypeEnum;
-
 import org.apache.commons.lang3.tuple.Pair;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
@@ -11,7 +10,7 @@ import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import java.util.List;
 
 public interface AnnotationTaskBlockRepository
-    extends JpaRepository<AnnotationTaskBlock, Integer>,
+    extends JpaRepository<AnnotationTaskBlock, Long>,
         JpaSpecificationExecutor<AnnotationTaskBlock> {
 
   AnnotationTaskBlock getOneByAnnotationTypeEqualsAndTextEquals(
@@ -24,7 +23,7 @@ public interface AnnotationTaskBlockRepository
    * @return 所有属于某个Task的Block
    */
   List<AnnotationTaskBlock> findByAnnotationTypeEqualsAndStateInAndTaskDocs_TaskDoc_Task_IdEquals(
-      AnnotationTypeEnum annotationType, List<AnnotationTaskState> states, int taskId);
+      AnnotationTypeEnum annotationType, List<AnnotationTaskState> states, long taskId);
 
   default Pair<AnnotationTaskBlock, Boolean> getOrCreateBlock(
       final AnnotationTypeEnum annotationType, final String text) {
@@ -36,5 +35,5 @@ public interface AnnotationTaskBlockRepository
   }
 
   List<AnnotationTaskBlock> findAllByStateInAndIdIn(
-      List<AnnotationTaskState> stateList, List<Integer> ids);
+      List<AnnotationTaskState> stateList, List<Long> ids);
 }
