+++ src/main/java/cn/malgo/annotation/dao/OriginalDocRepository.java
@@ -5,4 +5,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
 public interface OriginalDocRepository
-    extends JpaRepository<OriginalDoc, Integer>, JpaSpecificationExecutor<OriginalDoc> {}
+    extends JpaRepository<OriginalDoc, Long>, JpaSpecificationExecutor<OriginalDoc> {}
