+++ src/main/java/cn/malgo/annotation/request/GetDocByTaskRequest.java
@@ -6,7 +6,5 @@ import lombok.Data;
 @Data
 @AllArgsConstructor
 public class GetDocByTaskRequest {
-
-  int taskId;
-
+  long taskId;
 }
