+++ src/main/java/cn/malgo/annotation/request/LogOutRequest.java
@@ -2,8 +2,7 @@ package cn.malgo.annotation.request;
 
 import lombok.Data;
 
-/** Created by cjl on 2018/6/4. */
 @Data
 public class LogOutRequest {
-  private int userId;
+  private long userId;
 }
