+++ src/main/java/cn/malgo/annotation/request/SearchAnnotationRequest.java
@@ -5,7 +5,7 @@ import lombok.Value;
 @Value
 public class SearchAnnotationRequest {
   private final int annotationType;
-  private final int taskId;
+  private final long taskId;
 
   // term and type are both regex
   private final String term;
