+++ src/main/java/cn/malgo/annotation/request/SetUserStateRequest.java
@@ -4,11 +4,10 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-/** Created by cjl on 2018/5/30. */
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 public class SetUserStateRequest {
-  private int userId;
+  private long userId;
   private String currentState;
 }
