+++ src/main/java/cn/malgo/annotation/request/brat/AddRelationRequest.java
@@ -4,12 +4,11 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-/** Created by cjl on 2018/5/31. */
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 public class AddRelationRequest implements BaseAnnotationRequest {
-  private int id;
+  private long id;
   private String sourceTag;
   private String targetTag;
   private String relation;
