+++ src/main/java/cn/malgo/annotation/request/brat/DeleteRelationRequest.java
@@ -4,11 +4,10 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-/** Created by cjl on 2018/5/31. */
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 public class DeleteRelationRequest implements BaseAnnotationRequest {
-  private int id;
+  private long id;
   private String reTag;
 }
