+++ src/main/java/cn/malgo/annotation/request/brat/GetAutoAnnotationRequest.java
@@ -4,10 +4,9 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-/** Created by cjl on 2018/6/8. */
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 public class GetAutoAnnotationRequest {
-  private int id;
+  private long id;
 }
