+++ src/main/java/cn/malgo/annotation/request/brat/UpdateAnnotationRequest.java
@@ -4,12 +4,11 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-/** Created by cjl on 2018/5/31. */
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 public class UpdateAnnotationRequest implements BaseAnnotationRequest {
-  private int id;
+  private long id;
   private String tag;
   private String newType;
   private String autoAnnotation;
