+++ src/main/java/cn/malgo/annotation/request/task/ListAnnotationTaskRequest.java
@@ -8,7 +8,7 @@ import lombok.Value;
 public class ListAnnotationTaskRequest {
   int pageIndex;
   int pageSize;
-  int taskId;
+  long taskId;
   String name;
   List<AnnotationTaskState> taskStates;
 }
