+++ src/main/java/cn/malgo/annotation/service/AnnotationFixLogService.java
@@ -5,5 +5,5 @@ import cn.malgo.annotation.enums.AnnotationFixLogStateEnum;
 
 public interface AnnotationFixLogService {
   AnnotationFixLog insertOrUpdate(
-      int annotationId, int start, int end, AnnotationFixLogStateEnum state);
+      long annotationId, int start, int end, AnnotationFixLogStateEnum state);
 }
