+++ src/main/java/cn/malgo/annotation/vo/OriginalDocVO.java
@@ -6,7 +6,7 @@ import lombok.Value;
 @Value
 @AllArgsConstructor
 public class OriginalDocVO {
-  private int id; // 文档ID
+  private long id; // 文档ID
   private String type; // 文档类型
   private String state; // 文档状态
 }
