+++ src/main/java/cn/malgo/annotation/vo/ResetBlockToAnnotationResponse.java
@@ -6,5 +6,5 @@ import java.util.List;
 
 @Value
 public class ResetBlockToAnnotationResponse {
-  private final List<Integer> createdAnnotations;
+  private final List<Long> createdAnnotations;
 }
